Add tables with material UI

+++ src/helpersAdmin/deleteSeccion.js
@@ -2,11 +2,14 @@ import React from 'react'
 
 export const deleteSeccion = async(id) => {
    console.log("id",id);
+   if(!localStorage.getItem("token")) return alert ("registrate")
+    const {token} = JSON.parse(localStorage.getItem("token"));
      let config ={
         method:"DELETE",
         headers:{
         'Accept': 'application/json',
         'Content-Type': 'application/json',
+        'Authorization': `${token}`
         },
      }
    try{
